Conflicts:
	src/fitutil/SystematicManager.cpp
	src/fitutil/SystematicManager.h
	src/teststat/BinnedNLLH.cpp
	src/teststat/BinnedNLLH.h

+++ src/teststat/StatisticSum.h
@@ -18,7 +18,7 @@ class StatisticSum : public TestStatistic{
     virtual double Evaluate();
     virtual void   SetParameters(const ParameterDict& params_);
     virtual ParameterDict GetParameters() const;
-    virtual int    GetParameterCount() const;
+    virtual size_t GetParameterCount() const;
 
     virtual std::set<std::string> GetParameterNames() const;
